Rename: VoucherList -> VoucherType 로 Enum 클래스 이름변경
