+++ src/main/resources/logback.xml
@@ -1,6 +1,6 @@
 <configuration>
 
-    <property name="LOG_PATTERN" value ="%d{HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n" />
+    <property name="LOG_PATTERN" value="%d{HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n"/>
 
     <appender name="ROLLING_FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
         <file>logs/access.log</file>
@@ -13,6 +13,6 @@
     </appender>
 
     <root level="error">
-        <appender-ref ref="ROLLING_FILE" />
+        <appender-ref ref="ROLLING_FILE"/>
     </root>
 </configuration>
