+++ src/test/java/com/programmers/voucher/repository/HashMapVoucherRepositoryTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import java.util.Map;
 import java.util.UUID;
 
-import static com.programmers.voucher.voucher.VoucherList.FixedAmount;
+import static com.programmers.voucher.voucher.VoucherType.FixedAmount;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertSame;
 
@@ -36,7 +36,7 @@ class HashMapVoucherRepositoryTest {
 
         assertThat(store.size()).isEqualTo(0);
 
-        Voucher voucher1 = FixedAmount.createVoucher( 3000);
+        Voucher voucher1 = FixedAmount.createVoucher(3000);
         repository.registerVoucher(voucher1);
 
         assertThat(store.size()).isEqualTo(1);
@@ -49,7 +49,7 @@ class HashMapVoucherRepositoryTest {
 
     @Test
     void 맵에서Id로바우처조회() {
-        Voucher voucher = FixedAmount.createVoucher( 100000L);
+        Voucher voucher = FixedAmount.createVoucher(100000L);
         repository.registerVoucher(voucher);
 
         Voucher findOne = store.get(voucher.getVoucherId());
@@ -63,7 +63,7 @@ class HashMapVoucherRepositoryTest {
         assertThat(store.size()).isEqualTo(0);
 
         Voucher voucher1 = FixedAmount.createVoucher(3000);
-        Voucher voucher2 = FixedAmount.createVoucher( 3000);
+        Voucher voucher2 = FixedAmount.createVoucher(3000);
 
         repository.registerVoucher(voucher1);
         assertThat(store.size()).isEqualTo(1);
