+++ src/test/java/com/programmers/voucher/voucher/VoucherValidatorTest.java
@@ -3,8 +3,8 @@ package com.programmers.voucher.voucher;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.programmers.voucher.voucher.VoucherList.FixedAmount;
-import static com.programmers.voucher.voucher.VoucherList.PercentDiscount;
+import static com.programmers.voucher.voucher.VoucherType.FixedAmount;
+import static com.programmers.voucher.voucher.VoucherType.PercentDiscount;
 import static com.programmers.voucher.voucher.VoucherValidator.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
